ARS/backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
python-multipart
python-jose[cryptography]
httpx
pydantic

pymupdf
langchain
langchain-community
langchain-huggingface
langchain-text-splitters
langchain-groq
langchain-chroma
chromadb
rank-bm25
deep-translator
gTTS